change: refactored command handling a bit
