+++ src/main.rs
@@ -5,7 +5,6 @@ extern crate core;
 
 use crate::modules::utils;
 use anyhow::{anyhow, Result};
-use clap::{arg, Command, AppSettings};
 use std::process::exit;
 
 #[tokio::main]
@@ -18,53 +17,8 @@ async fn main() -> Result<()> {
 }
 
 async fn run() -> Result<()> {
-    let app = Command::new("bob")
-        .arg_required_else_help(true)
-        .subcommand(
-            Command::new("use")
-                .arg(arg!([VERSION]).required(true))
-                .about("Switch to a different neovim version"),
-        )
-        .subcommand(Command::new("ls").about("List all downloaded and installed versions"))
-        .get_matches();
-
-    if let Some(subcommand) = app.subcommand_matches("use") {
-        if let Err(error) = modules::use_handler::start(subcommand).await {
-            return Err(anyhow!(error));
-        }
-    }
-
-    if let Some(_) = app.subcommand_matches("ls") {
-        let downloads_dir = match utils::get_downloads_folder().await {
-            Ok(value) => value,
-            Err(error) => return Err(anyhow!(error)),
-        };
-
-        let paths = std::fs::read_dir(downloads_dir)?;
-        let mut versions = String::new();
-        let installed_version = match utils::get_current_version().await {
-            Some(value) => value,
-            None => return Err(anyhow!("Neovim is not installed")),
-        };
-        let mut first = true;
-        for path in paths {
-            let path_name = path.unwrap().file_name();
-            if !first {
-                if path_name.to_str().unwrap().contains(&installed_version) {
-                    versions += &format!(", {}(installed)", path_name.to_str().unwrap());
-                } else {
-                    versions += &format!(", {}", path_name.to_str().unwrap());
-                }
-            } else {
-                if path_name.to_str().unwrap().contains(&installed_version) {
-                    versions += &format!("{}(installed)", path_name.to_str().unwrap());
-                } else {
-                    versions += &format!("{}", path_name.to_str().unwrap());
-                }
-                first = false;
-            }
-        }
-        println!("{versions}");
+    if let Err(error) = modules::cli::start().await {
+        return Err(anyhow!(error));
     }
     Ok(())
 }
