+++ src/modules/ls_handler.rs
@@ -1,0 +1,36 @@
+use super::utils;
+use anyhow::{anyhow, Result};
+
+pub async fn start() -> Result<()> {
+    let downloads_dir = match utils::get_downloads_folder().await {
+        Ok(value) => value,
+        Err(error) => return Err(anyhow!(error)),
+    };
+
+    let paths = std::fs::read_dir(downloads_dir)?;
+    let mut versions = String::new();
+    let installed_version = match utils::get_current_version().await {
+        Some(value) => value,
+        None => return Err(anyhow!("Neovim is not installed")),
+    };
+    let mut first = true;
+    for path in paths {
+        let path_name = path.unwrap().file_name();
+        if !first {
+            if path_name.to_str().unwrap().contains(&installed_version) {
+                versions += &format!(", {}(installed)", path_name.to_str().unwrap());
+            } else {
+                versions += &format!(", {}", path_name.to_str().unwrap());
+            }
+        } else {
+            if path_name.to_str().unwrap().contains(&installed_version) {
+                versions += &format!("{}(installed)", path_name.to_str().unwrap());
+            } else {
+                versions += &format!("{}", path_name.to_str().unwrap());
+            }
+            first = false;
+        }
+    }
+    println!("{versions}");
+    Ok(())
+}
