+++ src/modules/mod.rs
@@ -1,3 +1,5 @@
+pub mod cli;
 mod expand_archive;
+pub mod ls_handler;
 pub mod use_handler;
 pub mod utils;
