+++ src/modules/utils.rs
@@ -100,7 +100,7 @@ pub async fn get_current_version() -> Option<String> {
     };
     let output = String::from_utf8_lossy(&*output.stdout).to_string();
     if output.contains("dev") {
-        return Some(String::from("nightly"))
+        return Some(String::from("nightly"));
     }
     let regex = Regex::new(r"v[0-9]\.[0-9]\.[0-9]").unwrap();
     Some(regex.find(output.as_str()).unwrap().as_str().to_owned())
